Subindo etapa final...

+++ projeto/application/views/telas/retrieve_categoria.php
@@ -1,5 +1,4 @@
 <?php
-if ($this->session->userdata('logado') == true) {
 // falta o teste da sessão...
 echo "<br>";
 echo "<div class='container-fluid'>";
@@ -14,17 +13,21 @@ echo 					"<div class='row'>";
 foreach ($categorias as $categoria) {
 echo 					"<div class='col-md-4'>";
 echo 					"<div class='card'>";
+echo 					"<a class='link-produtos' href='/CRUD_Produto/retrieve_categoria/?id=". $categoria->id. "'>";
 echo 						"<img class='card-img-top' src='../uploads/categorias/" . $categoria->id . ".jpg' alt='Card image cap'>";
 echo   						"<div class='card-body'>";
 echo  		   					"<h5 class='card-title text-center'>" . $categoria->nome . "</h5>";
 echo  		   						"<p class='card-text text-justify'>" . $categoria->descricao . "</p>";
+
+if ($this->session->userdata('logado') == true) {
 echo 							"<div class='text-right'>";
 echo 							"<a href='/CRUD_Categoria/update/?id=". $categoria->id ."'><img src='../images/icons/edit1_gray.png'>";
 echo 							"</img></a>";
 echo 							"<a class='excluir' href='/CRUD_Categoria/delete/?id=". $categoria->id ."'><img src='../images/icons/delete1_gray.png'>";
 echo 							"</img></a>";
 echo 							"</div>";
-echo 					"</div></div>";
+}
+echo 					"</a></div></div>";
 echo 				"</div>";
 }
 echo 				"</div>";
@@ -32,7 +35,6 @@ echo 				"</div>";
 echo 			"</div>";
 echo 		"</div>";
 echo "</div>";
-}
 
 ?>
